Force removal of lock if it originates from the same server
Sometimes, there is an error at the removal of the lock. The lock removes and prevents further processing.
With this patch, the server can detect it is the one which created the lock in the first place.

Thus it decides to remove it and process the new task. This should allow for remaining locks to be properly removed by the owning server in the time, preventing locking issues.