Catch RuntimeException when processing a task
Most of the time the RuntimeException is generated by infinispan not able to acquire or release a lock due to resource exhaustion on any of the servers being part of the infinispan cluster.
With this patch, this type or error will be catched, preventing the death of the associated Thread.

In addition, the error is logged for further investigation, and the abandonned task is put again at the end of the queue, allowing for another try.

Potential risk: A task causing issues systematically would be processed over again and again, which could cause an excessive overhead on the servers. This should be carefully monitored.